requirements.txt
Django==2.2.3
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
